cleaned up dontDestroy script so colors carry to board and menus work
